+++ retz-server/src/test/java/io/github/retz/jmx/MBeanCollectorTest.java
@@ -19,7 +19,6 @@ package io.github.retz.jmx;
 import com.j256.simplejmx.client.JmxClient;
 import com.j256.simplejmx.server.JmxServer;
 import io.github.retz.db.Database;
-import io.github.retz.scheduler.AdminConsole;
 import io.github.retz.scheduler.Launcher;
 import io.github.retz.scheduler.ServerConfiguration;
 import org.junit.After;
@@ -28,8 +27,6 @@ import org.junit.Test;
 
 import javax.management.ObjectName;
 import java.io.InputStream;
-import java.lang.reflect.Method;
-import java.util.List;
 import java.util.Optional;
 
 import static org.junit.Assert.*;
@@ -44,9 +41,7 @@ public class MBeanCollectorTest {
 
         ServerConfiguration config = new ServerConfiguration(in);
         Database.getInstance().init(config);
-        Method startJmxServer = AdminConsole.class.getDeclaredMethod("startJmxServer", ServerConfiguration.class, List.class);
-        startJmxServer.setAccessible(true);
-        Optional<JmxServer> server = (Optional<JmxServer>)startJmxServer.invoke(AdminConsole.class, config, MBeanCollector.beans());
+        Optional<JmxServer> server = RetzJmxServer.startJmxServer(config, MBeanCollector.beans());
         port = config.getJmxPort();
         jmxServer = server.get();
     }
